Finsihed part 1
Finally

+++ dec7/Sources/DirectoryContents/DeviceFile.swift
@@ -2,9 +2,9 @@ import Foundation
 
 class DeviceFile {
     private(set) var name: String
-    private(set) var size: uint32
+    private(set) var size: uint64
 
-    init(name: String, size: uint32) {
+    init(name: String, size: uint64) {
         self.name   = name
         self.size   = size
     }
